requirements.txt
bcrypt==4.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
enum-compat==0.0.3
filemod==2.1
future==0.18.2
idna==3.4
paramiko==2.11.0
pycparser==2.21
PyJWT==2.5.0
PyNaCl==1.5.0
pytz==2022.4
requests==2.28.1
six==1.16.0
urllib3==1.26.12
zope.interface==5.4.0
